+++ lib/Bet.dart
@@ -1,0 +1,9 @@
+class Bet {
+  final String title;
+  final String description;
+  final DateTime date;
+  final String initiator;
+  final String target;
+
+  Bet(this.title, this.description, this.date, this.initiator, this.target);
+}
